Refactored Application into seperate classes.
